Revert "vmgs: dont upgrade vmgs encryption on auto (#2280)" (#2377)
This reverts commit 64c47a90f0a48ca153e3277e6a65eb145db40a44. Due to a
performance issue that is not yet understood.

+++ vm/vmgs/vmgs_format/src/lib.rs
@@ -220,5 +220,5 @@ open_enum! {
 pub struct VmgsMarkers {
     pub reprovisioned: bool,
     #[bits(15)]
-    _reserved: u16,
+    _reserved: u32,
 }
